+++ src/logic/Arrays/searching/objectfind.ts
@@ -124,3 +124,5 @@ export function doesObjectExistInArray(sourceArray: any[], objectProperty : stri
   return result;
 
 }
+
+