+++ CHANGELOG.md
@@ -15,6 +15,7 @@ NOTE:  TURN BACK ON STRICT NULL CHECKS - turned off for PropPaneCols
 - remove unneccessary console.logs
 - fix a couple check4Gulp imports that pull in pnp2
 - update PinMeGroup to only auto-expand on PageInfo for now
+- add getNextIndex function from FadeCarousel component to src\logic\Arrays\searching\getNextIndex.ts
 
 ## 1.0.79 - 2022-Dec-23: fetch attacments pnpv2 update
 update fetch attachments from issue:  https://github.com/mikezimm/drilldown7/issues/302